Completed tasks #64, #65, #66, #67, #68, and #69.
Created the header as a separate fragment of the system, following the design from the wireframe (#64).
Positioned the organization logo in the left corner, as specified in the wireframe (#65).
Added a section header containing the name of the organization, following the wireframe (#66).
Implemented a link from the organization name text in the header to the organization landing page (#67).
Created the 'Scan me' section as described in the wireframe (#68).
Integrated a library or API that enables access to the device's camera for scanning purposes (#69).
All tasks have been successfully completed.