Completed tasks #73, #74, #75, #76 and #77.
Created the footer as a separate fragment of the system, following the design from the wireframe (#73).
Positioned the organization logo in the left corner, as specified in the wireframe (#74).
Added a section footer containing the name of the organization, following the wireframe (#75).
Implemented a link from the organization name text in the footer to the organization landing page (#76).
Implemented a link from the organization website text in the footer to the organization website page (#76).
All tasks have been successfully completed.